+++ lambdas/pkgpush/tests/test_index.py
@@ -4,19 +4,16 @@ import hashlib
 import io
 import json
 import unittest
-from http import HTTPStatus
 from unittest import mock
 
 import boto3
 import pytest
 from botocore.stub import Stubber
-from flask import Response
 
 import t4_lambda_pkgpush
 from quilt3.backends import get_package_registry
 from quilt3.packages import Package, PackageEntry
 from quilt3.util import PhysicalKey
-from t4_lambda_shared.decorator import Request
 
 
 def hash_data(data):
@@ -28,9 +25,9 @@ calculate_sha256_patcher = functools.partial(mock.patch, 'quilt3.packages.calcul
 
 class PackagePromoteTestBase(unittest.TestCase):
     credentials = {
-        'access_key': mock.sentinel.TEST_ACCESS_KEY,
-        'secret_key': mock.sentinel.TEST_SECRET_KEY,
-        'session_token': mock.sentinel.TEST_SESSION_TOKEN,
+        'aws_access_key_id': 'test_aws_access_key_id',
+        'aws_secret_access_key': 'test_aws_secret_access_key',
+        'aws_session_token': 'test_aws_session_token',
     }
     handler = staticmethod(t4_lambda_pkgpush.promote_package)
     parent_bucket = 'parent-bucket'
@@ -171,32 +168,22 @@ class PackagePromoteTestBase(unittest.TestCase):
              mock.patch('t4_lambda_pkgpush.get_package_registry', side_effect=side_effect, wraps=get_package_registry):
             yield
 
-    @classmethod
-    def _make_event(cls, body, credentials):
+    def make_request_wrapper(self, params, *, credentials, **kwargs):
         return {
-            'httpMethod': 'POST',
-            'path': '/foo',
-            'pathParameters': {},
-            'queryStringParameters': credentials,
-            'headers': None,
-            'body': body,
-            'isBase64Encoded': False,
+            "params": params,
+            "credentials": credentials,
         }
 
-    def make_request_wrapper(self, params, *, credentials, **kwargs):
-        return Request(
-            self._make_event(json.dumps(params), credentials=credentials),
+    def make_request_base(self, data, *, credentials, **kwargs):
+        result = self.handler(
+            self.make_request_wrapper(data, credentials=credentials, **kwargs),
             None,
         )
 
-    def make_request_base(self, data, *, credentials, **kwargs):
-        # This is a function before it gets wrapped with @api decorator.
-        # FIXME: find a cleaner way for this.
-        status, body, headers = self.handler.__wrapped__(
-            self.make_request_wrapper(data, credentials=credentials, **kwargs)
-        )
-        # Wrap in Flask response to ease migration from/to Flask.
-        return Response(body, status, headers)
+        # Check that result can be serialized to JSON.
+        json.dumps(result)
+
+        return result
 
     @mock.patch('time.time', mock.MagicMock(return_value=mock_timestamp))
     def make_request(self, params, **kwargs):
@@ -206,9 +193,7 @@ class PackagePromoteTestBase(unittest.TestCase):
             response = self.make_request_base(params, credentials=self.credentials, **kwargs)
 
         self.get_user_boto_session_mock.assert_called_once_with(
-            aws_access_key_id=mock.sentinel.TEST_ACCESS_KEY,
-            aws_secret_access_key=mock.sentinel.TEST_SECRET_KEY,
-            aws_session_token=mock.sentinel.TEST_SESSION_TOKEN,
+            **self.credentials,
         )
         reset_session_id_mock.assert_called_once_with()
 
@@ -351,20 +336,15 @@ class PackagePromoteTest(PackagePromoteTestBase):
 
                 with self.mock_successors({self.dst_registry: config_params}):
                     response = self.make_request(params)
-                    assert (response.status_code, response.json) == (
-                        200,
-                        {
-                            'status': 200,
+                    assert response == {
+                        "result": {
                             'top_hash': top_hash,
                         },
-                    )
+                    }
 
     def test_no_auth(self):
         resp = self.make_request_base({}, credentials={})
-        assert (resp.status_code, resp.data) == (
-            HTTPStatus.BAD_REQUEST,
-            b'{"message": "access_key, secret_key, session_token are required."}'
-        )
+        assert resp["error"]["name"] == "InvalidCredentials"
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_dst_is_not_successor(self):
@@ -380,12 +360,14 @@ class PackagePromoteTest(PackagePromoteTestBase):
             with self.subTest(successors=successors):
                 with self.mock_successors(successors):
                     response = self.make_request(params)
-                    assert (response.status_code, response.json) == (
-                        400,
-                        {
-                            'message': f'{self.dst_registry} is not configured as successor.',
+                    assert response == {
+                        "error": {
+                            "name": "InvalidSuccessor",
+                            "context": {
+                                "successor": self.dst_registry,
+                            }
                         },
-                    )
+                    }
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_invalid_successor(self):
@@ -401,12 +383,14 @@ class PackagePromoteTest(PackagePromoteTestBase):
             with self.subTest(registry_url=registry_url):
                 with self.mock_successors({registry_url: {}}):
                     response = self.make_request(params)
-                    assert (response.status_code, response.json) == (
-                        400,
-                        {
-                            'message': f'{registry_url} is not a valid S3 package registry.',
+                    assert response == {
+                        "error": {
+                            "name": "InvalidRegistry",
+                            "context": {
+                                "registry_url": registry_url,
+                            }
                         },
-                    )
+                    }
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_files_exceeded(self):
@@ -420,16 +404,15 @@ class PackagePromoteTest(PackagePromoteTestBase):
         with self.mock_successors({self.dst_registry: {'copy_data': True}}), \
              mock.patch(f't4_lambda_pkgpush.{self.max_files_const}', 1):
             response = self.make_request(params)
-            msg = (
-                f"Package has {self.files_number} files, "
-                f"but max supported number with `copy_data: true` is 1"
-            )
-            assert (response.status_code, response.json) == (
-                400,
-                {
-                    'message': msg,
+            assert response == {
+                "error": {
+                    "name": "TooManyFilesToCopy",
+                    "context": {
+                        "max_files": 1,
+                        "num_files": 2,
+                    }
                 },
-            )
+            }
 
     @mock.patch('t4_lambda_pkgpush.PROMOTE_PKG_MAX_MANIFEST_SIZE', 1)
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
@@ -453,13 +436,15 @@ class PackagePromoteTest(PackagePromoteTestBase):
 
             with self.mock_successors({self.dst_registry: {'copy_data': copy_data}}):
                 response = self.make_request(params)
-                msg = 'Manifest size of 42 exceeds supported limit of 1'
-                assert (response.status_code, response.json) == (
-                    400,
-                    {
-                        'message': msg,
+                assert response == {
+                    "error": {
+                        "name": "ManifestTooLarge",
+                        "context": {
+                            "max_size": 1,
+                            "size": 42,
+                        }
                     },
-                )
+                }
 
 
 @mock.patch('t4_lambda_pkgpush.PROMOTE_PKG_MAX_PKG_SIZE', 1)
@@ -478,16 +463,15 @@ class PackagePromoteTestSizeExceeded(PackagePromoteTestBase):
 
         with self.mock_successors({self.dst_registry: {'copy_data': True}}):
             response = self.make_request(params)
-            msg = (
-                f"Total package size is {self.file_size}, "
-                f"but max supported size with `copy_data: true` is 1"
-            )
-            assert (response.status_code, response.json) == (
-                400,
-                {
-                    'message': msg,
+            assert response == {
+                "error": {
+                    "name": "PackageTooLargeToCopy",
+                    "context": {
+                        "max_size": 1,
+                        "size": self.file_size,
+                    }
                 },
-            )
+            }
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_no_copy_data(self):
@@ -501,13 +485,11 @@ class PackagePromoteTestSizeExceeded(PackagePromoteTestBase):
 
         with self.mock_successors({self.dst_registry: {'copy_data': False}}):
             response = self.make_request(params)
-            assert (response.status_code, response.json) == (
-                200,
-                {
-                    'status': 200,
-                    'top_hash': top_hash,
+            assert response == {
+                "result": {
+                    "top_hash": top_hash,
                 },
-            )
+            }
 
 
 class PackageFromFolderTest(PackagePromoteTest):
@@ -767,7 +749,7 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
                         self.params,
                         entry,
                     ])
-                    assert pkg_response.status_code == 200
+                    assert "result" in pkg_response
 
     def test_object_level_meta(self):
         entry1 = self.gen_pkg_entry(
@@ -793,7 +775,7 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
                 self.params,
                 *entries,
             ])
-            assert pkg_response.status_code == 200
+            assert "result" in pkg_response
 
     def test_workflow_param(self):
         for params, expected_workflow in (
@@ -807,25 +789,37 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
                         params,
                         *self.package_entries,
                     ])
-                    assert pkg_response.status_code == 200
+                    assert "result" in pkg_response
 
     def test_invalid_parameters(self):
-        for params, expected_msg in (
+        for params, expected_error in (
             (
                 self.gen_params(name=...),
-                "'name' is a required property",
+                {
+                    "name": "InvalidInputParameters",
+                    "context": {"details": "'name' is a required property"},
+                },
             ),
             (
                 self.gen_params(name='invalid package name'),
-                'Invalid package name: invalid package name.',
+                {
+                    "name": "QuiltException",
+                    "context": {"details": "Invalid package name: invalid package name."},
+                },
             ),
             (
                 self.gen_params(registry=...),
-                "'registry' is a required property",
+                {
+                    "name": "InvalidInputParameters",
+                    "context": {"details": "'registry' is a required property"},
+                },
             ),
             (
                 self.gen_params(registry=self.dst_bucket),  # Not URL.
-                f'{self.dst_bucket} is not a valid S3 package registry.',
+                {
+                    "name": "InvalidRegistry",
+                    "context": {"registry_url": self.dst_bucket},
+                },
             ),
         ):
             with self.subTest(params=params):
@@ -833,8 +827,7 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
                     params,
                     *self.package_entries,
                 ])
-                assert pkg_response.status_code == 400
-                assert pkg_response.json['message'] == expected_msg
+                assert pkg_response["error"] == expected_error
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_invalid_entries_missing_required_props(self):
@@ -849,8 +842,12 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
                     },
                     entry,
                 ])
-                assert pkg_response.status_code == 400
-                assert pkg_response.get_json()['message'] == f"'{prop_name}' is a required property"
+                assert pkg_response == {
+                    "error": {
+                        "name": "InvalidInputParameters",
+                        "context": {"details": f"'{prop_name}' is a required property"},
+                    },
+                }
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_invalid_entries_non_url_pk(self):
@@ -862,8 +859,12 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
                 'physical_key': bad_pkey,
             },
         ])
-        assert pkg_response.status_code == 400
-        assert pkg_response.json['message'] == f'{bad_pkey} is not a valid s3 URL.'
+        assert pkg_response == {
+            "error": {
+                "name": "InvalidS3PhysicalKey",
+                "context": {"physical_key": bad_pkey},
+            },
+        }
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_invalid_entries_local_pk(self):
@@ -875,8 +876,12 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
                 'physical_key': bad_pkey,
             },
         ])
-        assert pkg_response.status_code == 400
-        assert pkg_response.json['message'] == f'{bad_pkey} is not in S3.'
+        assert pkg_response == {
+            "error": {
+                "name": "InvalidLocalPhysicalKey",
+                "context": {"physical_key": bad_pkey},
+            },
+        }
 
     @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
     def test_s3_error(self):
@@ -896,9 +901,16 @@ class PackageCreateNoHashingTestCase(PackageCreateTestCaseBase):
             self.params,
             *self.package_entries,
         ])
-        assert pkg_response.status_code == mock_http_code
-        assert mock_error_code in pkg_response.json['message']
-        assert mock_service_msg in pkg_response.json['message']
+        assert pkg_response == {
+            "error": {
+                "name": "AWSError",
+                "context": {
+                    "error_code": mock_error_code,
+                    "error_message": mock_service_msg,
+                    "status_code": mock_http_code,
+                },
+            },
+        }
 
 
 @mock.patch('quilt3.workflows.validate', lambda *args, **kwargs: None)
@@ -930,7 +942,7 @@ class PackageCreateWithHashingTestCase(PackageCreateTestCaseBase):
                         self.params,
                         entry,
                     ])
-                assert pkg_response.status_code == 200
+                assert "result" in pkg_response
 
 
 class HashCalculationTest(unittest.TestCase):
@@ -961,8 +973,9 @@ class HashCalculationTest(unittest.TestCase):
     @mock.patch.object(t4_lambda_pkgpush, 'S3_HASH_LAMBDA_MAX_FILE_SIZE_BYTES', 1)
     def test_calculate_pkg_hashes_too_large_file_error(self):
         s3_client = mock.MagicMock()
-        with pytest.raises(t4_lambda_pkgpush.FileTooLargeForHashing):
+        with pytest.raises(t4_lambda_pkgpush.PkgpushException) as excinfo:
             t4_lambda_pkgpush.calculate_pkg_hashes(s3_client, self.pkg)
+        assert excinfo.value.name == "FileTooLargeForHashing"
 
     def test_calculate_pkg_entry_hash(self):
         get_s3_client_mock = mock.MagicMock()
@@ -1027,6 +1040,7 @@ class HashCalculationTest(unittest.TestCase):
             },
         )
 
-        with pytest.raises(t4_lambda_pkgpush.S3HashLambdaUnhandledError):
+        with pytest.raises(t4_lambda_pkgpush.PkgpushException) as excinfo:
             t4_lambda_pkgpush.invoke_hash_lambda(test_url)
+        assert excinfo.value.name == "S3HashLambdaUnhandledError"
         lambda_client_stubber.assert_no_pending_responses()